Rename ThemeFeature to AbstracThemeFeature
